Add --strip support
In case the VCD file has a module in it, the --prefix
would only strip the first but not the second.

For example the VCD file could have:

dump.status[8191:0]
dump.wrapper_fibonacci.active

The --prefix "dump." would remove the first, but
we would still have the module (wrapper_fibonacci) inside of
the top module (dump), so adding code to deal with extra
stripping.

Signed-off-by: Konrad Rzeszutek Wilk <konrad@kernel.org>